Restructure project for typescript integration
