+++ server/package.json
@@ -11,7 +11,7 @@
     "test"
   ],
   "scripts": {
-    "start": "node index.js",
+    "start": "node src/index.js",
     "test": "mocha --exit"
   },
   "repository": {
