+++ server/src/config/example_dev.js
@@ -1,5 +1,6 @@
 // Install MongoDB locally for development: https://digipie.github.io/digidocs/mongodb/local-setup/
 // Add in the mongoURI key and rename the file as 'dev.js'
+// Find out more at: https://docs.mongodb.com/manual/reference/connection-string/
 module.exports = {
-  mongoURI: 'FILL THIS IN'
+  mongoURI: 'mongodb://localhost:27017'
 }
