+++ server/src/routes/productRoutes.js
